backend/requirements.txt
django==4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
